# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stock Broker Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
}

export interface User {
  id: number;
  email: string;
}

export interface Stock {
  ticker: string;
  name: string;
}

export type StockPriceTrend = typeof StockPriceTrend[keyof typeof StockPriceTrend];


export const StockPriceTrend = {
  up: 'up',
  down: 'down',
  neutral: 'neutral',
} as const;

export interface StockPrice {
  ticker: string;
  price: number;
  change: number;
  changePercent: number;
  trend: StockPriceTrend;
}

export interface Subscription {
  id: number;
  userId: number;
  ticker: string;
  createdAt: string;
}

export interface SubscriptionInput {
  ticker: string;
}

export interface PortfolioHolding {
  ticker: string;
  shares: number;
  avgCost: number;
  currentPrice: number;
  totalCost: number;
  totalValue: number;
  pnl: number;
  pnlPercent: number;
}

export interface Portfolio {
  holdings: PortfolioHolding[];
  totalValue: number;
  totalCost: number;
  totalPnl: number;
  totalPnlPercent: number;
}

export type TradeInputType = typeof TradeInputType[keyof typeof TradeInputType];


export const TradeInputType = {
  buy: 'buy',
  sell: 'sell',
} as const;

export interface TradeInput {
  ticker: string;
  type: TradeInputType;
  shares: number;
}

export interface TradeRecord {
  id: number;
  ticker: string;
  type: string;
  shares: number;
  price: number;
  total: number;
  executedAt: string;
}

export interface MessageResponse {
  message: string;
}