# lib/api-zod/src/generated/types/portfolio.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stock Broker Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { PortfolioHolding } from './portfolioHolding';

export interface Portfolio {
  holdings: PortfolioHolding[];
  totalValue: number;
  totalCost: number;
  totalPnl: number;
  totalPnlPercent: number;
}