# lib/api-zod/src/generated/types/portfolioHolding.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stock Broker Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface PortfolioHolding {
  ticker: string;
  shares: number;
  avgCost: number;
  currentPrice: number;
  totalCost: number;
  totalValue: number;
  pnl: number;
  pnlPercent: number;
}