# lib/api-zod/src/generated/types/stockPrice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stock Broker Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { StockPriceTrend } from './stockPriceTrend';

export interface StockPrice {
  ticker: string;
  price: number;
  change: number;
  changePercent: number;
  trend: StockPriceTrend;
}