# lib/api-zod/src/generated/types/subscription.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stock Broker Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Subscription {
  id: number;
  userId: number;
  ticker: string;
  createdAt: string;
}