# lib/api-zod/src/generated/types/tradeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stock Broker Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { TradeInputType } from './tradeInputType';

export interface TradeInput {
  ticker: string;
  type: TradeInputType;
  shares: number;
}