# lib/api-zod/src/generated/types/tradeRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Stock Broker Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface TradeRecord {
  id: number;
  ticker: string;
  type: string;
  shares: number;
  price: number;
  total: number;
  executedAt: string;
}